Move import resolution implementation to 'core'.
