+++ foundation/src/test/scala/quasar/matchers.scala
@@ -18,6 +18,7 @@ package quasar
 
 import slamdata.Predef._
 import quasar.RenderTree.ops._
+import quasar.fp.ski._
 import quasar.fp._
 
 import matryoshka._
@@ -45,3 +46,41 @@ trait TreeMatchers {
     }
   }
 }
+
+trait ValidationMatchers extends org.specs2.scalaz.ValidationMatchers {
+  def beEqualIfSuccess[E, A](expected: Validation[E, A]) =
+    new Matcher[Validation[E, A]] {
+      def apply[S <: Validation[E, A]](s: Expectable[S]) = {
+        val v = s.value
+
+        v.fold(
+          κ(result(
+            expected.fold(κ(true), κ(false)),
+            "both failed",
+            s"$v is not $expected",
+            s)),
+            a => expected.fold(
+              κ(result(false, "", "expected failure", s)),
+              ex => result(a == ex, "both are equal", s"$a is not $ex", s)))
+    }
+  }
+}
+
+trait DisjunctionMatchers extends org.specs2.scalaz.DisjunctionMatchers {
+  def beRightDisjOrDiff[A, B: Equal](expected: B)(implicit rb: RenderTree[B]): Matcher[A \/ B] = new Matcher[A \/ B] {
+    def apply[S <: A \/ B](s: Expectable[S]) = {
+      val v = s.value
+      v.fold(
+        a => result(false, s"$v is right", s"$v is not right", s),
+        b => {
+          val d = (b.render diff expected.render).shows
+          result(b ≟ expected,
+            s"\n$v is right and equals:\n$d",
+            s"\n$v is right but does not equal:\n$d",
+            s)
+        })
+    }
+  }
+}
+
+object QuasarMatchers extends ValidationMatchers with DisjunctionMatchers with TreeMatchers
