+++ sql/src/main/scala/quasar/sql/SQLParser.scala
@@ -17,7 +17,7 @@
 package quasar.sql
 
 import slamdata.Predef._
-import quasar.common.JoinType
+import quasar.common.{CIName, JoinType}
 import quasar.fp.ski._
 import quasar.fp._
