+++ sql/src/main/scala/quasar/sql/Sql.scala
@@ -19,13 +19,10 @@ package quasar.sql
 import slamdata.Predef._
 import quasar._, RenderTree.ops._
 import quasar.fp._
-import quasar.fp.ski._
-import quasar.frontend.CIName
+import quasar.common.CIName
 
 import matryoshka._
-import matryoshka.implicits._
 import monocle.macros.Lenses
-import pathy.Path._
 import scalaz._, Scalaz._
 import scalaz.Liskov._
 
@@ -185,29 +182,8 @@ object Sql {
   filter:      Option[A],
   groupBy:     Option[GroupBy[A]],
   orderBy:     Option[OrderBy[A]])
-    extends Sql[A] {
-  def substituteRelationVariable[M[_]: Monad, T](mapping: Vari[A] => M[A])(implicit
-    T0: Recursive.Aux[T, Sql],
-    T1: Corecursive.Aux[T, Sql],
-    ev: A <~< T
-  ): M[SemanticError \/ Select[A]] = {
-      val newRelation = relation.traverse(_.transformM[EitherT[M, SemanticError, ?], A]({
-        case VariRelationAST(vari, alias) =>
-          EitherT(mapping(vari).map(ev(_).project match {
-            case Ident(name) =>
-              posixCodec.parsePath(Some(_), Some(_), κ(None), κ(None))(name).cata(
-                TableRelationAST(_, alias).right,
-                SemanticError.GenericError(s"bad path: $name (note: absolute file path required)").left) // FIXME
-            // If the variable points to another variable, substitute the old one for the new one
-            case Vari(symbol) => VariRelationAST(Vari(symbol), alias).right
-            case x =>
-              SemanticError.GenericError(s"not a valid table name: ${pprint(x.embed)}").left // FIXME
-          }))
-        case otherRelation => otherRelation.point[EitherT[M, SemanticError, ?]]
-      }, _.point[EitherT[M, SemanticError, ?]]))
-      newRelation.map(r => this.copy(relation = r)).run
-  }
-}
+    extends Sql[A]
+
 @Lenses final case class Vari[A] private[sql] (symbol: String) extends Sql[A] {
   def map[B](f: A => B): Vari[B] = Vari(symbol)
 }
